start tracking built files usually its not a good idea to check built files into source control. however, starting at npm@5, prepublish scripts are no longer run automatically on npm install.
since we haven't actually published our indexing tweaks to npm, we need to make
sure the built output is immediately available in the branch for the benefit of
projects like esri-leaflet-vector.

someday i'll figure out how to rollup the dependency appropriately, but
for now the only way i could get it working was to falsely designate the built
output as an ES6 module via jsnext:main.

dont need mapbox-gl-dev.js